+++ engines/reevengi/movie/psx.h
@@ -57,10 +57,10 @@ public:
 
 	bool eos() const { return _srcStream->eos(); }
 
-	int32 pos() const { return _pos; }
-	int32 size() const { return _size; }
+	int64 pos() const { return _pos; }
+	int64 size() const { return _size; }
 
-	bool seek(int32 offs, int whence = SEEK_SET);
+	bool seek(int64 offs, int whence = SEEK_SET);
 };
 
 } // end of namespace Reevengi
